move only if logfile exists

+++ modules/install_steps.sh
@@ -498,7 +498,9 @@ finishing_cleanup() {
 }
 
 failure_cleanup() {
-    spawn "mv ${logfile} ${logfile}.failed" || warn "could not move ${logfile} to ${logfile}.failed"
+    if [ -f ${logfile} ]; then
+        spawn "mv ${logfile} ${logfile}.failed" || warn "could not move ${logfile} to ${logfile}.failed"
+    fi
     for mnt in $(awk '{ print $2; }' /proc/mounts | grep ^${chroot_dir} | sort -r | uniq); do
         spawn "umount ${mnt}" || warn "  could not unmount ${mnt}.. "
     done
